Redesigning the structure
I've been thinking deeply that maybe translating only the titles, topics, and comments are not so useful to people who can't understand English.

Although understanding English is essential. It make no sense to do a translation if the resource is in English and the titles in another language. So I propose to give alternatives resources to the translated language to approximate the result to the original given.

To make this process easier I will do some mock-ups. In the case of video lessons I would try to synthesize the information like a flashcard. (As the author did for remember the lessons).